refactor(shared): Split units.ts into domain-based modules for LLM optimization (#20)
Reorganize the 611-line units.ts file into a modular structure:

units/
├── index.ts    - UNITS aggregate + helpers + re-exports (177 lines)
├── types.ts    - UnitDefinition, UnitAbility, type aliases (77 lines)
├── constants.ts - UNIT_TYPE_*, RECRUIT_SITE_*, UNIT_ABILITY_* (42 lines)
├── ids.ts      - UnitId constants + type union (72 lines)
├── regular.ts  - Level 1 + Level 2 regular units (306 lines)
└── elite.ts    - Level 3 + Level 4 elite units (243 lines)

Benefits:
- Clear separation by concern: types, constants, IDs, definitions
- Unit definitions split by type (regular/elite) for focused editing
- Follows established pattern from enemies/ module refactor
- All exports preserved for backward compatibility

Updated imports in:
- actions.ts, index.ts, events/offers.ts, events/units.ts, types/clientState.ts
